release

+++ yt_downloader.py
@@ -2,7 +2,7 @@ import os
 
 from pytube import YouTube, Stream
 import shutil
-import moviepy.editor as editor
+from moviepy.editor import AudioFileClip, VideoFileClip
 
 
 class Downloader:
@@ -24,8 +24,11 @@ class Downloader:
     def __on_complete(stream: Stream, file_path):
         print('\nsuccessfully downloaded {file_type} from {title}'.format(file_type=stream.type, title=stream.title))
 
-    def download_video(self, resolution='360p', video_type='mp4', output_dir='./', video_name=None):
-        stream = self.yt.streams.filter(subtype='mp4', resolution=resolution, type='video').first()
+    def download_video(self, resolution='highest', video_type='mp4', output_dir='./', video_name=None):
+        if resolution == 'highest':
+            stream = self.yt.streams.filter(subtype='mp4', adaptive=True, type='video').first()
+        else:
+            stream = self.yt.streams.filter(subtype='mp4', resolution=resolution, type='video').first()
 
         if video_name is None:
             video_name = stream.default_filename
@@ -33,24 +36,35 @@ class Downloader:
             video_name = '{name}.{type}'.format(name=video_name, type=video_type)
 
         video_path = os.path.join(output_dir, video_name)
+        if os.path.exists(video_path):
+            return video_path, VideoFileClip(video_path).duration
+
+        # here we create the cache path and so the output path would be created at the same time
         cache_path = os.path.join(output_dir, 'cache/')
+        if not os.path.exists(cache_path):
+            os.makedirs(cache_path)
+
         temp_video = stream.download(output_path=cache_path, filename='temp_video.mp4')
+        video = VideoFileClip(temp_video)
+        duration = video.duration
 
         if stream.is_adaptive:
             temp_audio = self.yt.streams.filter(subtype='mp4', type='audio').first().download(output_path=cache_path)
 
             print('Now writing the audio into video... ')
-            video = editor.VideoFileClip(temp_video)
-            audio = editor.AudioFileClip(temp_audio)
+            audio = AudioFileClip(temp_audio)
             result_video = video.set_audio(audio)
             result_video.write_videofile(video_path)
 
             result_video.close()
-            video.close()
             audio.close()
-            shutil.rmtree(cache_path)
         else:
             os.rename(src=temp_video, dst=video_path)
+
+        video.close()
+
+        shutil.rmtree(cache_path)
+        return video_path
 
     def download_audio(self, audio_type='mp3', output_dir='./', audio_name=None):
         stream = self.yt.streams.filter(subtype='mp4', type='audio').first()
@@ -61,16 +75,25 @@ class Downloader:
 
         temp_video_path = stream.download(output_path=cache_path)
 
-        audio = editor.AudioFileClip(temp_video_path)
-
         if audio_name is None:
             audio_name = stream.default_filename.replace('mp4', audio_type)
         else:
             audio_name = '{name}.{type}'.format(name=audio_name, type=audio_type)
 
-        print('audio: ' + audio_name)
+        audio = AudioFileClip(temp_video_path)
+        duration = audio.duration
 
-        audio.write_audiofile(os.path.join(output_dir, audio_name))
+        audio_path = os.path.join(output_dir, audio_name)
+
+        # do not use ffmpeg codec 'libfdk_aac'
+        if audio_type == 'm4a':
+            audio.close()
+            os.rename(temp_video_path, audio_path)
+            return audio_path, duration
+
+        audio.write_audiofile(audio_path)
         audio.close()
         # clean cache
         shutil.rmtree(cache_path, ignore_errors=True)
+
+        return audio_path, duration
